Modificacion del tipo de firma de la query pasandole un decimal

+++ MyReserve/Models/Repository/RepositoryPeluqueria/IPeluqueria.cs
@@ -23,7 +23,7 @@ namespace MyReserve.Models.Repository.RepositoryPeluqueria {
         Task<IEnumerable<Categorias>> getCategorias();
         Task CrearServicios(Servicios servicio);
         Task<IEnumerable<Servicios>> getServiciosPeluqueriaCreados(int pelu_id);
-        Task actualizarServicioPeluqueria(int ser_id, string ser_nombre, int ser_precio, string ser_cat_id_fk);
+        Task actualizarServicioPeluqueria(int ser_id, string ser_nombre, decimal ser_precio, string ser_cat_id_fk);
         Task EliminarServicioPeluqueriaCreado(int ser_id);
     }
 }
